Use generics for channel merge function
