+++ iowrappers/utils.go
@@ -5,6 +5,8 @@ import (
 	"errors"
 	"github.com/modern-go/reflect2"
 	"github.com/weihesdlegend/Vacation-planner/POI"
+	"github.com/weihesdlegend/Vacation-planner/user"
+	"sync"
 )
 
 func scanRedisKeys(context context.Context, redisClient *RedisClient, redisKeyPrefix string) ([]string, error) {
@@ -40,4 +42,30 @@ func filter(places []POI.Place, condition func(place POI.Place) bool) []POI.Plac
 		}
 	}
 	return results
+}
+
+type View interface {
+	user.View | user.TravelPlanView
+}
+
+func merge[V View](cs ...chan V) chan V {
+	var wg sync.WaitGroup
+	out := make(chan V)
+
+	output := func(c <-chan V) {
+		for view := range c {
+			out <- view
+		}
+		wg.Done()
+	}
+	wg.Add(len(cs))
+	for _, c := range cs {
+		go output(c)
+	}
+
+	go func() {
+		wg.Wait()
+		close(out)
+	}()
+	return out
 }
